refactor: 카테고리 API 로직을 별도 서비스로 분리
